+++ Source/MainComponent.h
@@ -22,6 +22,7 @@ public:
     void drawComboBox(juce::Graphics&, int, int, bool, int, int, int, int, juce::ComboBox&) override;
     void drawButtonBackground(juce::Graphics&, juce::Button&, const juce::Colour&, bool, bool) override;
     void drawToggleButton(juce::Graphics&, juce::ToggleButton&, bool, bool) override;
+    juce::Label* createSliderTextBox(juce::Slider&) override;
 };
 
 class CircleEditor final : public juce::Component, public juce::SettableTooltipClient